@@ -1,4 +1,4 @@
-// automatically generated, do not modify
+// Code generated by the FlatBuffers compiler. DO NOT EDIT.
 
 package goserbench
 
@@ -10,9 +10,27 @@ type FlatBufferA struct {
 	_tab flatbuffers.Table
 }
 
+func GetRootAsFlatBufferA(buf []byte, offset flatbuffers.UOffsetT) *FlatBufferA {
+	n := flatbuffers.GetUOffsetT(buf[offset:])
+	x := &FlatBufferA{}
+	x.Init(buf, n+offset)
+	return x
+}
+
+func GetSizePrefixedRootAsFlatBufferA(buf []byte, offset flatbuffers.UOffsetT) *FlatBufferA {
+	n := flatbuffers.GetUOffsetT(buf[offset+flatbuffers.SizeUint32:])
+	x := &FlatBufferA{}
+	x.Init(buf, n+offset+flatbuffers.SizeUint32)
+	return x
+}
+
 func (rcv *FlatBufferA) Init(buf []byte, i flatbuffers.UOffsetT) {
 	rcv._tab.Bytes = buf
 	rcv._tab.Pos = i
+}
+
+func (rcv *FlatBufferA) Table() flatbuffers.Table {
+	return rcv._tab
 }
 
 func (rcv *FlatBufferA) Name() []byte {
@@ -31,6 +49,10 @@ func (rcv *FlatBufferA) BirthDay() int64 {
 	return 0
 }
 
+func (rcv *FlatBufferA) MutateBirthDay(n int64) bool {
+	return rcv._tab.MutateInt64Slot(6, n)
+}
+
 func (rcv *FlatBufferA) Phone() []byte {
 	o := flatbuffers.UOffsetT(rcv._tab.Offset(8))
 	if o != 0 {
@@ -47,12 +69,20 @@ func (rcv *FlatBufferA) Siblings() int32 {
 	return 0
 }
 
-func (rcv *FlatBufferA) Spouse() byte {
+func (rcv *FlatBufferA) MutateSiblings(n int32) bool {
+	return rcv._tab.MutateInt32Slot(10, n)
+}
+
+func (rcv *FlatBufferA) Spouse() bool {
 	o := flatbuffers.UOffsetT(rcv._tab.Offset(12))
 	if o != 0 {
-		return rcv._tab.GetByte(o + rcv._tab.Pos)
+		return rcv._tab.GetBool(o + rcv._tab.Pos)
 	}
-	return 0
+	return false
+}
+
+func (rcv *FlatBufferA) MutateSpouse(n bool) bool {
+	return rcv._tab.MutateBoolSlot(12, n)
 }
 
 func (rcv *FlatBufferA) Money() float64 {
@@ -60,10 +90,16 @@ func (rcv *FlatBufferA) Money() float64 {
 	if o != 0 {
 		return rcv._tab.GetFloat64(o + rcv._tab.Pos)
 	}
-	return 0
+	return 0.0
 }
 
-func FlatBufferAStart(builder *flatbuffers.Builder) { builder.StartObject(6) }
+func (rcv *FlatBufferA) MutateMoney(n float64) bool {
+	return rcv._tab.MutateFloat64Slot(14, n)
+}
+
+func FlatBufferAStart(builder *flatbuffers.Builder) {
+	builder.StartObject(6)
+}
 func FlatBufferAAddName(builder *flatbuffers.Builder, name flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(0, flatbuffers.UOffsetT(name), 0)
 }
@@ -76,10 +112,12 @@ func FlatBufferAAddPhone(builder *flatbuffers.Builder, phone flatbuffers.UOffset
 func FlatBufferAAddSiblings(builder *flatbuffers.Builder, siblings int32) {
 	builder.PrependInt32Slot(3, siblings, 0)
 }
-func FlatBufferAAddSpouse(builder *flatbuffers.Builder, spouse byte) {
-	builder.PrependByteSlot(4, spouse, 0)
+func FlatBufferAAddSpouse(builder *flatbuffers.Builder, spouse bool) {
+	builder.PrependBoolSlot(4, spouse, false)
 }
 func FlatBufferAAddMoney(builder *flatbuffers.Builder, money float64) {
-	builder.PrependFloat64Slot(5, money, 0)
+	builder.PrependFloat64Slot(5, money, 0.0)
 }
-func FlatBufferAEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT { return builder.EndObject() }
+func FlatBufferAEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {
+	return builder.EndObject()
+}
