@@ -11,6 +11,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -22,7 +23,7 @@ var _ = math.Inf
 // is compatible with the proto package it is being compiled against.
 // A compilation error at this line likely means your copy of the
 // proto package needs to be updated.
-const _ = proto.GoGoProtoPackageIsVersion2 // please upgrade the proto package
+const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
 type GogoProtoBufA struct {
 	Name     string  `protobuf:"bytes,1,req,name=name" json:"name"`
@@ -47,7 +48,7 @@ func (m *GogoProtoBufA) XXX_Marshal(b []byte, deterministic bool) ([]byte, error
 		return xxx_messageInfo_GogoProtoBufA.Marshal(b, m, deterministic)
 	} else {
 		b = b[:cap(b)]
-		n, err := m.MarshalTo(b)
+		n, err := m.MarshalToSizedBuffer(b)
 		if err != nil {
 			return nil, err
 		}
@@ -115,28 +116,31 @@ func init() {
 func init() { proto.RegisterFile("structdef-gogo.proto", fileDescriptor_445e4de7f4e06aee) }
 
 var fileDescriptor_445e4de7f4e06aee = []byte{
-	// 240 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x4c, 0x8c, 0x31, 0x4e, 0xc3, 0x30,
-	0x18, 0x85, 0x63, 0x37, 0xa9, 0x8a, 0x25, 0x16, 0x8b, 0xc1, 0xaa, 0x90, 0xb1, 0x98, 0xbc, 0xb4,
-	0x39, 0x03, 0x11, 0x12, 0x2b, 0xe2, 0x06, 0x75, 0x70, 0x9c, 0x48, 0xc4, 0x7f, 0x14, 0xdb, 0x43,
-	0x6f, 0xc1, 0x9d, 0x58, 0x3a, 0x76, 0x64, 0x42, 0x28, 0xb9, 0x08, 0x8a, 0x49, 0xa0, 0xe3, 0xfb,
-	0xde, 0xfb, 0x1e, 0xb9, 0x71, 0xbe, 0x0f, 0xa5, 0x7f, 0xd5, 0xd5, 0xce, 0x80, 0x81, 0x7d, 0xd7,
-	0x83, 0x07, 0x4a, 0x0c, 0x38, 0xdd, 0x2b, 0x6d, 0xcb, 0x7a, 0xbb, 0x33, 0x8d, 0xaf, 0x83, 0xda,
-	0x97, 0xd0, 0xe6, 0xd3, 0x24, 0x8f, 0x13, 0x15, 0xaa, 0x98, 0x62, 0xc8, 0xff, 0xd5, 0xfb, 0x0f,
-	0x44, 0xae, 0x9f, 0xc0, 0xc0, 0xf3, 0x94, 0x8a, 0x50, 0x3d, 0x50, 0x46, 0x52, 0x7b, 0x68, 0x35,
-	0x43, 0x02, 0xcb, 0xab, 0x22, 0x3d, 0x7d, 0xdd, 0x25, 0x2f, 0x91, 0x50, 0x41, 0x36, 0xaa, 0xe9,
-	0x7d, 0xfd, 0x78, 0x38, 0x32, 0x2c, 0xb0, 0x5c, 0xcd, 0xed, 0x1f, 0xa5, 0x5b, 0x92, 0x75, 0x35,
-	0x58, 0xcd, 0x56, 0x17, 0xf2, 0x2f, 0x9a, 0x6c, 0xd7, 0xa8, 0xb7, 0xc6, 0x1a, 0xc7, 0x52, 0x81,
-	0x65, 0xb6, 0xd8, 0x0b, 0xa5, 0xb7, 0x64, 0xed, 0x3a, 0x08, 0x4e, 0xb3, 0x4c, 0x60, 0xb9, 0x99,
-	0xfb, 0x99, 0x4d, 0xdf, 0x2d, 0x58, 0x7d, 0x64, 0x6b, 0x81, 0x25, 0x5a, 0xbe, 0x23, 0x2a, 0xd8,
-	0x69, 0xe0, 0xe8, 0x3c, 0x70, 0xf4, 0x3d, 0x70, 0xf4, 0x3e, 0xf2, 0xe4, 0x3c, 0xf2, 0xe4, 0x73,
-	0xe4, 0xc9, 0x4f, 0x00, 0x00, 0x00, 0xff, 0xff, 0x8e, 0x18, 0xeb, 0x76, 0x31, 0x01, 0x00, 0x00,
+	// 280 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x4c, 0x90, 0xbd, 0x4e, 0xc3, 0x30,
+	0x14, 0x85, 0x63, 0xf7, 0x47, 0xc5, 0x12, 0x4b, 0xc4, 0x60, 0x55, 0xc8, 0x58, 0x4c, 0x59, 0xda,
+	0x3c, 0x00, 0x2c, 0x54, 0x48, 0xac, 0x88, 0x81, 0x81, 0xa5, 0x72, 0x82, 0xeb, 0x58, 0x34, 0xbe,
+	0x91, 0xed, 0x0c, 0xe5, 0x29, 0x78, 0x27, 0x96, 0x8e, 0x1d, 0x99, 0x10, 0x4a, 0x5e, 0x04, 0xc5,
+	0x24, 0xb4, 0xe3, 0xf9, 0xee, 0xfd, 0xce, 0x70, 0xc8, 0x85, 0xf3, 0xb6, 0xce, 0xfd, 0xab, 0xdc,
+	0x2c, 0x14, 0x28, 0x58, 0x56, 0x16, 0x3c, 0xc4, 0x44, 0x81, 0x93, 0x36, 0x93, 0x26, 0x2f, 0xe6,
+	0x0b, 0xa5, 0x7d, 0x51, 0x67, 0xcb, 0x1c, 0xca, 0xb4, 0x7b, 0x49, 0xc3, 0x4b, 0x56, 0x6f, 0x42,
+	0x0a, 0x21, 0x3d, 0xaa, 0xd7, 0x9f, 0x88, 0x9c, 0x3f, 0x80, 0x82, 0xc7, 0x2e, 0xad, 0xea, 0xcd,
+	0x5d, 0x4c, 0xc9, 0xd8, 0x88, 0x52, 0x52, 0xc4, 0x71, 0x72, 0xb6, 0x1a, 0xef, 0xbf, 0xaf, 0xa2,
+	0xa7, 0x40, 0x62, 0x4e, 0x66, 0x99, 0xb6, 0xbe, 0xb8, 0x17, 0x3b, 0x8a, 0x39, 0x4e, 0x46, 0xfd,
+	0xf5, 0x9f, 0xc6, 0x73, 0x32, 0xa9, 0x0a, 0x30, 0x92, 0x8e, 0x4e, 0xe4, 0x3f, 0xd4, 0xd9, 0x4e,
+	0x67, 0x5b, 0x6d, 0x94, 0xa3, 0x63, 0x8e, 0x93, 0xc9, 0x60, 0x0f, 0x34, 0xbe, 0x24, 0x53, 0x57,
+	0x41, 0xed, 0x24, 0x9d, 0x70, 0x9c, 0xcc, 0xfa, 0x7b, 0xcf, 0xba, 0xee, 0x12, 0x8c, 0xdc, 0xd1,
+	0x29, 0xc7, 0x09, 0x1a, 0xba, 0x03, 0x5a, 0x3d, 0xef, 0x1b, 0x86, 0x0e, 0x0d, 0x43, 0x3f, 0x0d,
+	0x43, 0x1f, 0x2d, 0x8b, 0x0e, 0x2d, 0x8b, 0xbe, 0x5a, 0x16, 0xbd, 0xdc, 0x9e, 0xcc, 0x21, 0xb6,
+	0x32, 0xf7, 0x05, 0x94, 0xc2, 0xa5, 0x0a, 0xd6, 0x4e, 0x5a, 0x2d, 0xb6, 0xfa, 0x5d, 0x78, 0x0d,
+	0x66, 0x1d, 0xc6, 0x2b, 0x85, 0x7d, 0x73, 0x37, 0xc7, 0x31, 0x7f, 0x03, 0x00, 0x00, 0xff, 0xff,
+	0x52, 0xf5, 0x1a, 0x3d, 0x6f, 0x01, 0x00, 0x00,
 }
 
 func (m *GogoProtoBufA) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
-	n, err := m.MarshalTo(dAtA)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
 	if err != nil {
 		return nil, err
 	}
@@ -144,47 +148,56 @@ func (m *GogoProtoBufA) Marshal() (dAtA []byte, err error) {
 }
 
 func (m *GogoProtoBufA) MarshalTo(dAtA []byte) (int, error) {
-	var i int
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *GogoProtoBufA) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
 	_ = i
 	var l int
 	_ = l
-	dAtA[i] = 0xa
-	i++
-	i = encodeVarintStructdefGogo(dAtA, i, uint64(len(m.Name)))
-	i += copy(dAtA[i:], m.Name)
-	dAtA[i] = 0x10
-	i++
-	i = encodeVarintStructdefGogo(dAtA, i, uint64(m.BirthDay))
-	dAtA[i] = 0x1a
-	i++
-	i = encodeVarintStructdefGogo(dAtA, i, uint64(len(m.Phone)))
-	i += copy(dAtA[i:], m.Phone)
-	dAtA[i] = 0x20
-	i++
-	i = encodeVarintStructdefGogo(dAtA, i, uint64(m.Siblings))
-	dAtA[i] = 0x28
-	i++
+	i -= 8
+	encoding_binary.LittleEndian.PutUint64(dAtA[i:], uint64(math.Float64bits(float64(m.Money))))
+	i--
+	dAtA[i] = 0x31
+	i--
 	if m.Spouse {
 		dAtA[i] = 1
 	} else {
 		dAtA[i] = 0
 	}
-	i++
-	dAtA[i] = 0x31
-	i++
-	encoding_binary.LittleEndian.PutUint64(dAtA[i:], uint64(math.Float64bits(float64(m.Money))))
-	i += 8
-	return i, nil
+	i--
+	dAtA[i] = 0x28
+	i = encodeVarintStructdefGogo(dAtA, i, uint64(m.Siblings))
+	i--
+	dAtA[i] = 0x20
+	i -= len(m.Phone)
+	copy(dAtA[i:], m.Phone)
+	i = encodeVarintStructdefGogo(dAtA, i, uint64(len(m.Phone)))
+	i--
+	dAtA[i] = 0x1a
+	i = encodeVarintStructdefGogo(dAtA, i, uint64(m.BirthDay))
+	i--
+	dAtA[i] = 0x10
+	i -= len(m.Name)
+	copy(dAtA[i:], m.Name)
+	i = encodeVarintStructdefGogo(dAtA, i, uint64(len(m.Name)))
+	i--
+	dAtA[i] = 0xa
+	return len(dAtA) - i, nil
 }
 
 func encodeVarintStructdefGogo(dAtA []byte, offset int, v uint64) int {
+	offset -= sovStructdefGogo(v)
+	base := offset
 	for v >= 1<<7 {
 		dAtA[offset] = uint8(v&0x7f | 0x80)
 		v >>= 7
 		offset++
 	}
 	dAtA[offset] = uint8(v)
-	return offset + 1
+	return base
 }
 func (m *GogoProtoBufA) Size() (n int) {
 	if m == nil {
@@ -204,14 +217,7 @@ func (m *GogoProtoBufA) Size() (n int) {
 }
 
 func sovStructdefGogo(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozStructdefGogo(x uint64) (n int) {
 	return sovStructdefGogo(uint64((x << 1) ^ uint64((int64(x) >> 63))))
@@ -391,10 +397,7 @@ func (m *GogoProtoBufA) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthStructdefGogo
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthStructdefGogo
 			}
 			if (iNdEx + skippy) > l {
@@ -430,6 +433,7 @@ func (m *GogoProtoBufA) Unmarshal(dAtA []byte) error {
 func skipStructdefGogo(dAtA []byte) (n int, err error) {
 	l := len(dAtA)
 	iNdEx := 0
+	depth := 0
 	for iNdEx < l {
 		var wire uint64
 		for shift := uint(0); ; shift += 7 {
@@ -461,10 +465,8 @@ func skipStructdefGogo(dAtA []byte) (n int, err error) {
 					break
 				}
 			}
-			return iNdEx, nil
 		case 1:
 			iNdEx += 8
-			return iNdEx, nil
 		case 2:
 			var length int
 			for shift := uint(0); ; shift += 7 {
@@ -485,55 +487,30 @@ func skipStructdefGogo(dAtA []byte) (n int, err error) {
 				return 0, ErrInvalidLengthStructdefGogo
 			}
 			iNdEx += length
-			if iNdEx < 0 {
-				return 0, ErrInvalidLengthStructdefGogo
-			}
-			return iNdEx, nil
 		case 3:
-			for {
-				var innerWire uint64
-				var start int = iNdEx
-				for shift := uint(0); ; shift += 7 {
-					if shift >= 64 {
-						return 0, ErrIntOverflowStructdefGogo
-					}
-					if iNdEx >= l {
-						return 0, io.ErrUnexpectedEOF
-					}
-					b := dAtA[iNdEx]
-					iNdEx++
-					innerWire |= (uint64(b) & 0x7F) << shift
-					if b < 0x80 {
-						break
-					}
-				}
-				innerWireType := int(innerWire & 0x7)
-				if innerWireType == 4 {
-					break
-				}
-				next, err := skipStructdefGogo(dAtA[start:])
-				if err != nil {
-					return 0, err
-				}
-				iNdEx = start + next
-				if iNdEx < 0 {
-					return 0, ErrInvalidLengthStructdefGogo
-				}
-			}
-			return iNdEx, nil
+			depth++
 		case 4:
-			return iNdEx, nil
+			if depth == 0 {
+				return 0, ErrUnexpectedEndOfGroupStructdefGogo
+			}
+			depth--
 		case 5:
 			iNdEx += 4
-			return iNdEx, nil
 		default:
 			return 0, fmt.Errorf("proto: illegal wireType %d", wireType)
 		}
+		if iNdEx < 0 {
+			return 0, ErrInvalidLengthStructdefGogo
+		}
+		if depth == 0 {
+			return iNdEx, nil
+		}
 	}
-	panic("unreachable")
+	return 0, io.ErrUnexpectedEOF
 }
 
 var (
-	ErrInvalidLengthStructdefGogo = fmt.Errorf("proto: negative length found during unmarshaling")
-	ErrIntOverflowStructdefGogo   = fmt.Errorf("proto: integer overflow")
+	ErrInvalidLengthStructdefGogo        = fmt.Errorf("proto: negative length found during unmarshaling")
+	ErrIntOverflowStructdefGogo          = fmt.Errorf("proto: integer overflow")
+	ErrUnexpectedEndOfGroupStructdefGogo = fmt.Errorf("proto: unexpected end of group")
 )
