@@ -1,121 +1,192 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.28.1
+// 	protoc        v3.21.12
 // source: structdef.proto
 
 package goserbench
 
 import (
-	fmt "fmt"
-	proto "github.com/golang/protobuf/proto"
-	math "math"
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
 )
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
-
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
 type ProtoBufA struct {
-	Name                 *string  `protobuf:"bytes,1,req,name=name" json:"name,omitempty"`
-	BirthDay             *int64   `protobuf:"varint,2,req,name=birthDay" json:"birthDay,omitempty"`
-	Phone                *string  `protobuf:"bytes,3,req,name=phone" json:"phone,omitempty"`
-	Siblings             *int32   `protobuf:"varint,4,req,name=siblings" json:"siblings,omitempty"`
-	Spouse               *bool    `protobuf:"varint,5,req,name=spouse" json:"spouse,omitempty"`
-	Money                *float64 `protobuf:"fixed64,6,req,name=money" json:"money,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Name     *string  `protobuf:"bytes,1,req,name=name" json:"name,omitempty"`
+	BirthDay *int64   `protobuf:"varint,2,req,name=birthDay" json:"birthDay,omitempty"`
+	Phone    *string  `protobuf:"bytes,3,req,name=phone" json:"phone,omitempty"`
+	Siblings *int32   `protobuf:"varint,4,req,name=siblings" json:"siblings,omitempty"`
+	Spouse   *bool    `protobuf:"varint,5,req,name=spouse" json:"spouse,omitempty"`
+	Money    *float64 `protobuf:"fixed64,6,req,name=money" json:"money,omitempty"`
 }
 
-func (m *ProtoBufA) Reset()         { *m = ProtoBufA{} }
-func (m *ProtoBufA) String() string { return proto.CompactTextString(m) }
-func (*ProtoBufA) ProtoMessage()    {}
+func (x *ProtoBufA) Reset() {
+	*x = ProtoBufA{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_structdef_proto_msgTypes[0]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ProtoBufA) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ProtoBufA) ProtoMessage() {}
+
+func (x *ProtoBufA) ProtoReflect() protoreflect.Message {
+	mi := &file_structdef_proto_msgTypes[0]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ProtoBufA.ProtoReflect.Descriptor instead.
 func (*ProtoBufA) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3829f236e6ac66c8, []int{0}
+	return file_structdef_proto_rawDescGZIP(), []int{0}
 }
 
-func (m *ProtoBufA) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_ProtoBufA.Unmarshal(m, b)
-}
-func (m *ProtoBufA) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_ProtoBufA.Marshal(b, m, deterministic)
-}
-func (m *ProtoBufA) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_ProtoBufA.Merge(m, src)
-}
-func (m *ProtoBufA) XXX_Size() int {
-	return xxx_messageInfo_ProtoBufA.Size(m)
-}
-func (m *ProtoBufA) XXX_DiscardUnknown() {
-	xxx_messageInfo_ProtoBufA.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_ProtoBufA proto.InternalMessageInfo
-
-func (m *ProtoBufA) GetName() string {
-	if m != nil && m.Name != nil {
-		return *m.Name
+func (x *ProtoBufA) GetName() string {
+	if x != nil && x.Name != nil {
+		return *x.Name
 	}
 	return ""
 }
 
-func (m *ProtoBufA) GetBirthDay() int64 {
-	if m != nil && m.BirthDay != nil {
-		return *m.BirthDay
+func (x *ProtoBufA) GetBirthDay() int64 {
+	if x != nil && x.BirthDay != nil {
+		return *x.BirthDay
 	}
 	return 0
 }
 
-func (m *ProtoBufA) GetPhone() string {
-	if m != nil && m.Phone != nil {
-		return *m.Phone
+func (x *ProtoBufA) GetPhone() string {
+	if x != nil && x.Phone != nil {
+		return *x.Phone
 	}
 	return ""
 }
 
-func (m *ProtoBufA) GetSiblings() int32 {
-	if m != nil && m.Siblings != nil {
-		return *m.Siblings
+func (x *ProtoBufA) GetSiblings() int32 {
+	if x != nil && x.Siblings != nil {
+		return *x.Siblings
 	}
 	return 0
 }
 
-func (m *ProtoBufA) GetSpouse() bool {
-	if m != nil && m.Spouse != nil {
-		return *m.Spouse
+func (x *ProtoBufA) GetSpouse() bool {
+	if x != nil && x.Spouse != nil {
+		return *x.Spouse
 	}
 	return false
 }
 
-func (m *ProtoBufA) GetMoney() float64 {
-	if m != nil && m.Money != nil {
-		return *m.Money
+func (x *ProtoBufA) GetMoney() float64 {
+	if x != nil && x.Money != nil {
+		return *x.Money
 	}
 	return 0
 }
 
-func init() {
-	proto.RegisterType((*ProtoBufA)(nil), "goserbench.ProtoBufA")
+var File_structdef_proto protoreflect.FileDescriptor
+
+var file_structdef_proto_rawDesc = []byte{
+	0x0a, 0x0f, 0x73, 0x74, 0x72, 0x75, 0x63, 0x74, 0x64, 0x65, 0x66, 0x2e, 0x70, 0x72, 0x6f, 0x74,
+	0x6f, 0x12, 0x0a, 0x67, 0x6f, 0x73, 0x65, 0x72, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x22, 0x9b, 0x01,
+	0x0a, 0x09, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x42, 0x75, 0x66, 0x41, 0x12, 0x12, 0x0a, 0x04, 0x6e,
+	0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x02, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12,
+	0x1a, 0x0a, 0x08, 0x62, 0x69, 0x72, 0x74, 0x68, 0x44, 0x61, 0x79, 0x18, 0x02, 0x20, 0x02, 0x28,
+	0x03, 0x52, 0x08, 0x62, 0x69, 0x72, 0x74, 0x68, 0x44, 0x61, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x70,
+	0x68, 0x6f, 0x6e, 0x65, 0x18, 0x03, 0x20, 0x02, 0x28, 0x09, 0x52, 0x05, 0x70, 0x68, 0x6f, 0x6e,
+	0x65, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x69, 0x62, 0x6c, 0x69, 0x6e, 0x67, 0x73, 0x18, 0x04, 0x20,
+	0x02, 0x28, 0x05, 0x52, 0x08, 0x73, 0x69, 0x62, 0x6c, 0x69, 0x6e, 0x67, 0x73, 0x12, 0x16, 0x0a,
+	0x06, 0x73, 0x70, 0x6f, 0x75, 0x73, 0x65, 0x18, 0x05, 0x20, 0x02, 0x28, 0x08, 0x52, 0x06, 0x73,
+	0x70, 0x6f, 0x75, 0x73, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x6d, 0x6f, 0x6e, 0x65, 0x79, 0x18, 0x06,
+	0x20, 0x02, 0x28, 0x01, 0x52, 0x05, 0x6d, 0x6f, 0x6e, 0x65, 0x79, 0x42, 0x3e, 0x5a, 0x3c, 0x67,
+	0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x61, 0x6c, 0x65, 0x63, 0x74, 0x68,
+	0x6f, 0x6d, 0x61, 0x73, 0x2f, 0x67, 0x6f, 0x5f, 0x73, 0x65, 0x72, 0x69, 0x61, 0x6c, 0x69, 0x7a,
+	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x5f, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x73,
+	0x3b, 0x67, 0x6f, 0x73, 0x65, 0x72, 0x62, 0x65, 0x6e, 0x63, 0x68,
 }
 
-func init() { proto.RegisterFile("structdef.proto", fileDescriptor_3829f236e6ac66c8) }
+var (
+	file_structdef_proto_rawDescOnce sync.Once
+	file_structdef_proto_rawDescData = file_structdef_proto_rawDesc
+)
 
-var fileDescriptor_3829f236e6ac66c8 = []byte{
-	// 169 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x3c, 0xcc, 0x4d, 0xca, 0x83, 0x30,
-	0x10, 0xc6, 0x71, 0x8c, 0x1f, 0xe8, 0x6c, 0x5e, 0x08, 0x2f, 0x25, 0x74, 0x15, 0xba, 0xca, 0xaa,
-	0x77, 0x68, 0xe9, 0x01, 0x4a, 0x6e, 0xa0, 0x76, 0xfc, 0x80, 0x9a, 0x91, 0x4c, 0x5c, 0x78, 0x96,
-	0x5e, 0xb6, 0x44, 0x8b, 0xbb, 0xf9, 0x0d, 0xfc, 0x1f, 0xf8, 0xe3, 0xe0, 0x97, 0x36, 0xbc, 0xb0,
-	0xbb, 0xce, 0x9e, 0x02, 0x49, 0xe8, 0x89, 0xd1, 0x37, 0xe8, 0xda, 0xe1, 0xf2, 0x49, 0xa0, 0x7a,
-	0xc6, 0xef, 0x7d, 0xe9, 0x6e, 0x52, 0x42, 0xe6, 0xea, 0x09, 0x55, 0xa2, 0x85, 0xa9, 0xec, 0x76,
-	0xcb, 0x33, 0x94, 0xcd, 0xe8, 0xc3, 0xf0, 0xa8, 0x57, 0x25, 0xb4, 0x30, 0xa9, 0x3d, 0x2c, 0xff,
-	0x21, 0x9f, 0x07, 0x72, 0xa8, 0xd2, 0x2d, 0xd8, 0x11, 0x0b, 0x1e, 0x9b, 0xf7, 0xe8, 0x7a, 0x56,
-	0x99, 0x16, 0x26, 0xb7, 0x87, 0xe5, 0x09, 0x0a, 0x9e, 0x69, 0x61, 0x54, 0xb9, 0x16, 0xa6, 0xb4,
-	0x3f, 0xc5, 0xa5, 0x89, 0x1c, 0xae, 0xaa, 0xd0, 0xc2, 0x24, 0x76, 0xc7, 0x37, 0x00, 0x00, 0xff,
-	0xff, 0x57, 0x66, 0x93, 0x38, 0xbb, 0x00, 0x00, 0x00,
+func file_structdef_proto_rawDescGZIP() []byte {
+	file_structdef_proto_rawDescOnce.Do(func() {
+		file_structdef_proto_rawDescData = protoimpl.X.CompressGZIP(file_structdef_proto_rawDescData)
+	})
+	return file_structdef_proto_rawDescData
+}
+
+var file_structdef_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
+var file_structdef_proto_goTypes = []interface{}{
+	(*ProtoBufA)(nil), // 0: goserbench.ProtoBufA
+}
+var file_structdef_proto_depIdxs = []int32{
+	0, // [0:0] is the sub-list for method output_type
+	0, // [0:0] is the sub-list for method input_type
+	0, // [0:0] is the sub-list for extension type_name
+	0, // [0:0] is the sub-list for extension extendee
+	0, // [0:0] is the sub-list for field type_name
+}
+
+func init() { file_structdef_proto_init() }
+func file_structdef_proto_init() {
+	if File_structdef_proto != nil {
+		return
+	}
+	if !protoimpl.UnsafeEnabled {
+		file_structdef_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*ProtoBufA); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+	}
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: file_structdef_proto_rawDesc,
+			NumEnums:      0,
+			NumMessages:   1,
+			NumExtensions: 0,
+			NumServices:   0,
+		},
+		GoTypes:           file_structdef_proto_goTypes,
+		DependencyIndexes: file_structdef_proto_depIdxs,
+		MessageInfos:      file_structdef_proto_msgTypes,
+	}.Build()
+	File_structdef_proto = out.File
+	file_structdef_proto_rawDesc = nil
+	file_structdef_proto_goTypes = nil
+	file_structdef_proto_depIdxs = nil
 }
